refactor: Move getBoundingBox to api folder
