+++ user_interface/src/common/api/getBoundingBox.js
